src/pydantic-ai-agent-tools/pyproject.toml
[project]
name = "pydantic-ai-agent-tools"
version = "0.1.0"
description = "Agent with calculator, DuckDuckGo search, and Open-Meteo weather tools"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.0.0",
    "httpx>=0.27",
    "numexpr>=2.14.1",
    "pydantic-ai>=1.7.0",
    "python-dotenv>=1.1.0",
]
